# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mixd",
    "description": "Personal music metadata hub",
    "version": "0.10.0"
  },
  "paths": {
...
    "/api/v1/workflows/{workflow_id}/run": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Run Workflow Endpoint",
        "description": "Start a workflow execution. Returns immediately with operation_id + run_id.",
        "operationId": "run_workflow_endpoint_api_v1_workflows__workflow_id__run_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunStartedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/{workflow_id}/runs": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "List Workflow Runs",
        "description": "List execution history for a workflow.",
        "operationId": "list_workflow_runs_api_v1_workflows__workflow_id__runs_get",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workflow Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_WorkflowRunSummarySchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/{workflow_id}/runs/{run_id}": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "Get Workflow Run",
        "description": "Get a single run with node execution details.",
        "operationId": "get_workflow_run_api_v1_workflows__workflow_id__runs__run_id__get",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workflow Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunDetailSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/operation-runs": {
      "get": {
        "tags": [
          "operation-runs"
        ],
        "summary": "List Operation Runs",
        "description": "List the user's audit-log rows newest-first, keyset-paginated.\n\n``type=imports`` (default) restricts to import/sync/apply runs that\nsurface in the Import History UI. ``type=all`` returns every row.\n``status=running`` powers operation-awareness: the in-flight rows the\nfrontend re-attaches to (each carries its ``operation_id`` SSE handle).",
        "operationId": "list_operation_runs_api_v1_operation_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "imports",
              "title": "Type"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OperationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationRunListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/operation-runs/{run_id}": {
      "get": {
        "tags": [
          "operation-runs"
        ],
        "summary": "Get Operation Run",
        "description": "Get one run with the full issues payload.\n\nReturns 404 for both not-found AND not-owner \u2014 deliberately uniform\nso the response can't be used to infer the existence of another\nuser's run.",
        "operationId": "get_operation_run_api_v1_operation_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationRunDetailSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/operation-runs/{run_id}/retry-failed": {
      "post": {
        "tags": [
          "operation-runs"
        ],
        "summary": "Retry Failed Operation",
        "description": "Re-run only the failed items of a terminal import run.\n\nServer-reconstructed: the failed connector-playlist ids come from the run's\n``issues`` and the connector + direction from its ``request_params``. The\nowner is taken from auth, never from stored data. Re-invokes the *same*\nimport use case with the failed subset as a fresh ``OperationRun`` \u2014 no new\norchestration. Returns 409 when there's nothing retryable (so the caller can\nfall back to \"View log\").",
        "operationId": "retry_failed_operation_api_v1_operation_runs__run_id__retry_failed_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationStartedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JsonDict": {
        "additionalProperties": {
          "$ref": "#/components/schemas/JsonValue-Output"
        },
        "type": "object"
      },
...
      "OperationRunDetailSchema": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "operation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Id"
          },
          "operation_type": {
            "type": "string",
            "title": "Operation Type"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "complete",
              "error",
              "cancelled"
            ],
            "title": "Status"
          },
          "counts": {
            "$ref": "#/components/schemas/JsonDict"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/JsonDict"
            },
            "type": "array",
            "title": "Issues"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          },
          "initiated_by": {
            "type": "string",
            "title": "Initiated By"
          }
        },
        "type": "object",
        "required": [
          "id",
          "operation_id",
          "operation_type",
          "started_at",
          "ended_at",
          "status",
          "counts",
          "issues",
          "retryable",
          "initiated_by"
        ],
        "title": "OperationRunDetailSchema",
        "description": "Full audit-log row including the issues array."
      },
      "OperationRunListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/OperationRunSummarySchema"
            },
            "type": "array",
            "title": "Data"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "data",
          "limit",
          "next_cursor"
        ],
        "title": "OperationRunListResponse",
        "description": "List shape: data array plus opaque next-page cursor."
      },
      "OperationRunSummarySchema": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "operation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Id"
          },
          "operation_type": {
            "type": "string",
            "title": "Operation Type"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "complete",
              "error",
              "cancelled"
            ],
            "title": "Status"
          },
          "counts": {
            "$ref": "#/components/schemas/JsonDict"
          },
          "issue_count": {
            "type": "integer",
            "title": "Issue Count"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          },
          "initiated_by": {
            "type": "string",
            "title": "Initiated By"
          }
        },
        "type": "object",
        "required": [
          "id",
          "operation_id",
          "operation_type",
          "started_at",
          "ended_at",
          "status",
          "counts",
          "issue_count",
          "retryable",
          "initiated_by"
        ],
        "title": "OperationRunSummarySchema",
        "description": "Lightweight row for the list view (no full ``issues`` payload).\n\nThe list endpoint returns this shape; full ``issues`` come back from\nthe per-run detail endpoint so a 100-issue run doesn't bloat the\nlist response."
      },
...
      "OperationStartedResponse": {
        "properties": {
          "operation_id": {
            "type": "string",
            "title": "Operation Id"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "type": "object",
        "required": [
          "operation_id"
        ],
        "title": "OperationStartedResponse",
        "description": "Returned immediately when a long-running operation is launched.\n\n``run_id`` is the persistent ``OperationRun`` audit-log row id, present\nwhen the route writes one via the seam-level recorder. The frontend\ncaches it alongside ``operation_id`` so the post-run toast can deep-\nlink to ``/settings/imports?run=<run_id>`` without a second round trip."
      },
      "OperationStatus": {
        "type": "string",
        "enum": [
          "running",
          "complete",
          "error",
          "cancelled"
        ]
      },
...
      "PaginatedResponse_WorkflowRunSummarySchema_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/WorkflowRunSummarySchema"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "data",
          "limit",
          "offset"
        ],
        "title": "PaginatedResponse[WorkflowRunSummarySchema]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkflowDefSchema-Output": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "version": {
            "type": "string",
            "title": "Version",
            "default": "1.0"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/WorkflowTaskDefSchema-Output"
            },
            "type": "array",
            "title": "Tasks",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "WorkflowDefSchema"
      },
...
      "WorkflowRunDetailSchema": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "workflow_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workflow Id"
          },
          "run_number": {
            "type": "integer",
            "title": "Run Number"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "completed",
              "failed",
              "cancelled",
              "crashed"
            ],
            "title": "Status"
          },
          "definition_version": {
            "type": "integer",
            "title": "Definition Version",
            "default": 1
          },
          "operation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Id"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "heartbeat_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heartbeat At"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "output_track_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Track Count"
          },
          "output_playlist_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Playlist Id"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "definition_snapshot": {
            "$ref": "#/components/schemas/WorkflowDefSchema-Output"
          },
          "output_tracks": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Output Tracks",
            "default": []
          },
          "metric_columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Metric Columns",
            "default": []
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/WorkflowRunNodeSchema"
            },
            "type": "array",
            "title": "Nodes",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "workflow_id",
          "run_number",
          "status",
          "definition_snapshot"
        ],
        "title": "WorkflowRunDetailSchema"
      },
...
      "WorkflowRunStartedResponse": {
        "properties": {
          "operation_id": {
            "type": "string",
            "title": "Operation Id"
          },
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          }
        },
        "type": "object",
        "required": [
          "operation_id",
          "run_id"
        ],
        "title": "WorkflowRunStartedResponse"
      },
      "WorkflowRunSummarySchema": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "workflow_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workflow Id"
          },
          "run_number": {
            "type": "integer",
            "title": "Run Number"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "completed",
              "failed",
              "cancelled",
              "crashed"
            ],
            "title": "Status"
          },
          "definition_version": {
            "type": "integer",
            "title": "Definition Version",
            "default": 1
          },
          "operation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Id"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "heartbeat_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heartbeat At"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "output_track_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Track Count"
          },
          "output_playlist_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Playlist Id"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workflow_id",
          "run_number",
          "status"
        ],
        "title": "WorkflowRunSummarySchema"
      },
...
    }
  }
}